refactor(register) add register request

+++ app/Http/Controllers/EventController.php
@@ -4,14 +4,27 @@ namespace App\Http\Controllers;
 
 
 use App\Models\Event;
+use GuzzleHttp\Middleware;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Support\Facades\Gate;
 
-class EventController extends Controller
+class EventController extends Controller implements HasMiddleware
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+
+    public static function middleware(): array
+    {
+        return[
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
+
+
+
+    public function index(): \Illuminate\Database\Eloquent\Collection
     {
         return Event::all();
     }
@@ -34,7 +47,7 @@ class EventController extends Controller
             'max_capacity' => 'nullable|integer',
         ]);
 
-        $event = Event::create($fields);
+        $event = $request->user()->create($fields);
 
         return $event;
 
@@ -53,6 +66,8 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
+
+        Gate::authorize('modify', $event);
         $fields = $request->validate([
             'title' => 'required|max:255',
             'description' => 'required',
@@ -76,6 +91,8 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        Gate::authorize('modify', $event);
+
         $event->delete();
         return ['message' => 'Event deleted'];
     }
